Refactoring: Moving I/O functionality into netplay_io.c
