+++ Makefile.common
@@ -1123,6 +1123,7 @@ ifeq ($(HAVE_NETWORKING), 1)
    # Netplay
 	DEFINES += -DHAVE_NETWORK_CMD
 	OBJ += network/netplay/netplay_frontend.o \
+			network/netplay/netplay_io.o \
 			network/netplay/netplay_net.o \
 			network/netplay/netplay_common.o \
 			network/netplay/netplay_discovery.o \
