+++ griffin/griffin.c
@@ -882,6 +882,7 @@ NETPLAY
 ============================================================ */
 #ifdef HAVE_NETWORKING
 #include "../network/netplay/netplay_frontend.c"
+#include "../network/netplay/netplay_io.c"
 #include "../network/netplay/netplay_net.c"
 #include "../network/netplay/netplay_common.c"
 #include "../network/netplay/netplay_discovery.c"
